requirements.txt
openai
python-dotenv
httpx
